// src/type.ts
export interface Car {
  _id: string;
  make: string;
  model: string;
  year: number;
  price: number;
  color: string;
  mileage: number;
  fuelType: string;
  transmission: string;
  condition: string;
  imageUrl: string;
  engineSize: string;
  horsePower: string;
  doors: number;
  seats: number;
}